fix(release): Linux 동반자 릴리스 검증을 실제 실행
Constraint: 범위 밖 회귀 금지

🐙 Autopus <noreply@autopus.co>
