+++ pkg/companionmanifest/release_darwin_producer_test.go
