+++ pkg/companionmanifest/release_tool_helpers_test.go
@@ -8,8 +8,6 @@ import (
 	"testing"
 )
 
-const acceptedNotaryID = "123e4567-e89b-42d3-a456-426614174000"
-
 func darwinReleaseToolEnv(t *testing.T, dir string) []string {
 	t.Helper()
 	tools := filepath.Join(dir, "tools")
